fix: restore build after legacy bindings + threading refactor
- Re-add missing legacy URL.origin getter shims (window/worker)
- Use ScriptSourceOverrideFetcher for module graph fetch overrides (external_script_sources is now Arc<Mutex<_>>)
- Fix flex measure-cache key test to match Option<u64> bits

+++ src/layout/contexts/flex.rs
@@ -15745,9 +15745,9 @@ mod tests {
       AvailableSpace::Definite(w) => assert_eq!(w, 434.0),
       other => panic!("expected snapped available width to remain definite, got {other:?}"),
     }
-
+ 
     // Ensure the key itself reflects the preserved width.
-    assert_eq!(key.0, Some(super::f32_to_canonical_bits(434.0)));
+    assert_eq!(key.0, Some(super::f32_to_canonical_bits(434.0) as u64));
   }
 
   #[test]
